📝 Temporarily changes SSH key comment
This should be changed back to variable usage because Ansible _should_ install a new SSH key to GitHub for each installation. However, while setting up a new work laptop, this wasn't the case. I didn't have time to troubleshoot it, so I just changed this comment to an extant key on GitHub for now.